fixup! mdns: add new connection property.
The keyfile reader/writer generates the code based on the
GObject property. In this case, it can only handle the value
as integers (contrary to nmcli, to which you specify an enum
type nm_setting_connection_mdns_get_type()).

Contrary to nmcli, keyfile is not (primarily) for manual editing,
so it doesn't need to accept pretty names for the values. Integers
are good enough.

One day, we should refactor keyfile to also use
clients/common/nm-meta-setting-desc.c and thus also treat the
property not only as integer, but as enum. But that should be done
for all properties the same way, not by now implementing each enum
property by hand.