device: ppp: start stage3 only after we receive IP configuration
The pppd plugin first sends a phase change signal to RUNNING and then
sends IPv4 and IPv6 configurations. NM knows the actual interface name
(and thus populates the ifindex) only after an IP configuration is
received from the plugin.

If we start configuring IPv4 and IPv6 as soon as the PPP status is
RUNNING, it is possible to hit an assertion because
activate_stage3_ip_config_start() tries to capture the current IPv6
configuration when the device ifindex is not known yet.

Instead, let's start configuring IP only after the ifindex is known.

https://bugzilla.redhat.com/show_bug.cgi?id=1515829