fixup! device: restore original dynamic IPv4 configuration on reapply
I don't think it can happen that we every try to set the config
to the current value. And if we would, applied_config_init()
wouild handle that correctly (first refing the new config,
before clearing the current). Only change is that we still
would call ip4_config_merge_and_apply(), but that seems right
to me too.