core: use array instead of GSList for existing-connections in complete_connection()
Most of the time when GSList is used, it'd better not be.
It's not convenient to use, and requires a slice allocation for
each element in the list.

Especially, since the list is originally already present as
array. Avoid the clone of the list and adjust the API to
use the array instead of the GSList.