core: cleanup generating con-name and ifname for nm_utils_complete_generic()
A notable change is that the loops are not arbitrarily limited
to 10000 or 500. I think that is better, because the limit previously
was arbitrary choosen. Was it too high, or too low? Note that there is
no endless loop, because eventually we will find a candidate name that is
not yet in use (because we won't have UINT_MAX conflicting names).

Also, build a lookup index of the names. This ensures that the functions
have a worst case runtime of O(n), instead of O(n^2).