manager: avoid cloning list of existing-connections in impl_manager_add_and_activate_connection()
The @connections list is eventually passed to nm_utils_complete_generic().
At no place, we require a particular order of the elements, hence there
is no need to sort the list.

If we don't sort it, we also don't need a shallow-copy of the array,
we can directly use the cached list as returned by nm_settings_get_connections().
This avoids the additional copy, and quite likely it already has the list
cached.

Note, that the cached list becomes invalidated when a connection is added
or removed. But if that happens, also a shallow-copy of the list would
not be safe, and we would need to ref each element in the list -- which
we were not doing before. So, this isn't more fragile then before.
At most, you are more likely to encounter a crash, while before you
might not have noticed when the list was modified -- because
accidentally the connection entry wasn't unrefed yet, or you'd
already visited it. So yes, if the list is accidentally modified (a
bug), it's more likely to encounter a crash then before. But the bug
already existed.