core: don't localize strings for daemon
NetworkManager deamon does not run in the context of the
(client) user, hence it usually has no explicit locale set. Even if
a admin configures a locale for system daemons, it is not necessarily
the right locale in the client context.

Getting localization of strings right would be important. For example,
when a request fails, the error message should be in the language of
the client.
However, that is a lot of work to get right. Either the client has
to provide a localization context (so the server can generated messages
in the right language), or the client is given a method to localize
messages (e.g. by not relying on generating the message by the server,
but by generating an error code that can be looked up). It's
complicated, but just translating strings server side doesn't help.

Probably we should build libnm-core.a twice, once with l10n for libnm's
use case, and once without for the server.

The problem that the server doesn't know the l10n context is especially
bad for the name of the generated default-wired-connection, which is
something like "Wired Connection 1". See also related bug bgo#666516.