device: unify tracking of routes that can currently not be configured
Kernel refuses to let NetworkManager configure certain routes, if
certain conditions are not met:

1) the IPv6 route has a RTA_PREFSRC (src) that is still tentative.
This is a kernel limitation. Basically, kernel requires user space to
*wait* until the references pref-src is no longer tentative. Either
NetworkManager does the waiting, or it could only support pref-src
route attribute together with addresses that are marked as
IFA_F_OPTIMISTIC. NetworkManager already makes an effort to support
this.

2) the IPv4 or IPv6 route specifies a gateway that is not onlink
and the onlink attribute is not set for IPv4 (onlink attribute
is not supported for IPv6 routes). This happens for example in the
common case where the user configures static addresses with a
gateway received via DHCP. One could argue that the user cannot possibly
know which gateway will be assigned on this device, or it could be
argued, if he already knows the gateway, he is required to also configure
a onlink route for this gateway. However, NetworkManager again tries
to be helpful and support this. See for example commit
7070d17cedd09d07f12ce977dd1e16cecf8d4b45.

We shall track all these addresses the same way. Hence, unify the
tracking of addresses that can currently not be configured -- but maybe
later.