device: retry adding routes with a gateway that fail to be added
When the user configures a gateway route without a direct route to
the gateway, he might expect that we receive the missing route via
DHCP. We still need to wait and retry later to add this route, possibly
when we received the requirements from DHCP.

This is a horrible hack. Commit 7070d17cedd09d07f12ce977dd1e16cecf8d4b45
already tries to solve this differently. But now that we also call
ip_config_merge_and_apply() to handle link local addresses, we call it
more often and quite possibly before we received a DHCP configuration.

What I dislike the most, how we take the result of
ip_config_merge_and_apply() to determine whether to fail the IP
configuration. But we do this very inconsistently and probably wrong.
The only reason why this currently doesn't blow up, is because the only
case where ip_config_merge_and_apply() can fail, is when we fail to add
a manually configured route (NM_IP_CONFIG_SOURCE_USER).